instageo/new_apps/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.0.1
pydantic==2.10.4
redis==6.2.0
rq==2.4.0
rq-dashboard==0.8.3.2
types-redis>=4.6.0
absl-py==2.3.1
titiler.core==0.22.4
ray[serve]
neptune
ptflops
tomli
datashader
# Model-specific requirements
pytorch_lightning
torch
timm==1.0.15
einops
tensorboard
omegaconf
hydra-core
torchvision
huggingface_hub
